hkirk: hs_real uses StablePtr + keep Haskell runtime alive while an hs_real exists
This fixes the "I do not own this lock" error in GHC's runtimes of
version >= 8.2 when the Haskell runtime is deinitialized using
shutdownHaskellAndExit() in GHC-src rts/RtsStartup.c. This is the
case when GHC links a program without the option '-no-hs-main'
thereby using rts/RtsStartup.c:realMain() which calls the above
shutdown function which in turn (pre 8.2 did) *reset* the
hs_init_counter responsible for balancing hs_init() calls with
hs_exit() calls. This bug was present since Aug 2011 [1] and has
been fixed in [2] which is available in GHC since version 8.2.0.

A workaround for earlier versions is to roll an own main() in C
and avoid calling the buggy GHC-internal shutdownHaskellAndExit()
in favour of the standard hs_{init,exit}() pairs.

[1] <https://git.haskell.org/ghc.git/commit/d18b5d53e74318e4a6bc2ad0557ff71a00c1abe1>
[2] <https://git.haskell.org/ghc.git/commit/9cbcdb4863064753df0fff9054b7b7c6b3188b64>